+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Bugsnag Notifiers on other platforms.
 * Add `unhandledRejections` to `BugsnagErrorTypes`
   [#567](https://github.com/bugsnag/bugsnag-cocoa/pull/567)
 
+* Rename `Bugsnag` start methods
+  [#566](https://github.com/bugsnag/bugsnag-cocoa/pull/566)
+
 * Rename `OnSend` to `OnSendError`
   [#562](https://github.com/bugsnag/bugsnag-cocoa/pull/562)